… app/src/main/java/app/revanced/integrations/returnyoutubedislike/requests/ReturnYouTubeDislikeRoutes.java
… app/src/main/java/app/revanced/integrations/settingsmenu/ReturnYouTubeDislikeSettingsFragment.java
… app/src/main/java/app/revanced/integrations/sponsorblock/requests/SBRequester.java
… app/src/main/java/app/revanced/integrations/utils/LogHelper.java
… app/src/main/java/app/revanced/integrations/utils/ReVancedUtils.java