I changed the background thread pool so all threads run at max priority.

This change seems to reduce the average time needed to load the video UI, as the RYD vote fetch now gets priority over other background threads (the fetch is blocking the UI, it definitely should get priority). 

Currently this thread pool is only used by RYD for the fetch votes calls.