@LisoUseInAIKyrios 2022-12-12
Benchmarking is not precise since it's affected by whatever is going on in the background, how fast YouTube responds for each video, internet connection speed at that moment, etc, etc, etc.

Without patch.  Loading 20 random videos:

```
revanced: ReturnYouTubeDislike: UI thread paused for: 1071ms waiting for vote fetch to complete
revanced: ReturnYouTubeDislike: UI thread paused for: 1221ms waiting for vote fetch to complete
revanced: ReturnYouTubeDislike: UI thread paused for: 661ms waiting for vote fetch to complete
revanced: ReturnYouTubeDislike: UI thread paused for: 61ms waiting for vote fetch to complete
```

With patch.  Loading same 20 random videos:

```
revanced: ReturnYouTubeDislike: UI thread paused for: 762ms waiting for vote fetch to complete
revanced: ReturnYouTubeDislike: UI thread paused for: 159ms waiting for vote fetch to complete

```