@LisoUseInAIKyrios 2022-12-12
without the patch, the UI waited 4 out of 20 times.

With the patch, it waited 2 out of 20 times.