@oSumAtrIX 2022-12-12
Oh I see, i thought these were samples taken out of the 20.