@LisoUseInAIKyrios 2022-12-16
I added logging to the API network calls, and added user viewable statistics if debugging is enabled:
https://freeimage.host/i/HoDNhKB

Also added a Toast if a connection times out, or if a rate limit is encountered.